requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
charset-normalizer==3.0.1
colorama==0.4.6
decorator==5.1.1
discord.py==2.2.2
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
pip==22.3
psutil==5.9.4
pyte==0.8.1
setuptools==65.5.0
six==1.16.0
wcwidth==0.2.6
win-unicode-console==0.5
yarl==1.8.2

SQLAlchemy~=2.0.5.post1